tweak compiler warnings
git-svn-id: http://vtp.googlecode.com/svn/trunk@6697 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtdata/Building.h
@@ -86,7 +86,7 @@ public:
 
 	void Set(int doors, int windows, const char *material);
 	void AddFeature(int code, float width = -1.0f, float vf1 = 0.0f, float vf2 = 1.0f);
-	int NumFeatures() const { return m_Features.size(); }
+	size_t NumFeatures() const { return m_Features.size(); }
 	int NumFeaturesOfCode(int code);
 	float FixedFeaturesWidth();
 	float ProportionTotal();
